Authenticated User added

+++ src/main/java/com/team/updevic001/services/impl/CommentServiceImpl.java
@@ -10,11 +10,13 @@ import com.team.updevic001.exceptions.ResourceNotFoundException;
 import com.team.updevic001.model.dtos.request.CommentDto;
 import com.team.updevic001.model.dtos.response.comment.ResponseCommentDto;
 import com.team.updevic001.services.interfaces.CommentService;
+import com.team.updevic001.utility.AuthHelper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -26,10 +28,11 @@ public class CommentServiceImpl implements CommentService {
     private final AdminServiceImpl adminServiceImpl;
     private final LessonServiceImpl lessonServiceImpl;
     private final CourseServiceImpl courseServiceImpl;
+    private final AuthHelper authHelper;
 
     @Override
-    public void addCommentToCourse(String userId, String courseId, CommentDto commentDto) {
-        User user = adminServiceImpl.findUserById(userId);
+    public void addCommentToCourse(String courseId, CommentDto commentDto) {
+        User user = authHelper.getAuthenticatedUser();
         Course course = courseServiceImpl.findCourseById(courseId);
         Comment comment = Comment.builder()
                 .content(commentDto.getContent())
@@ -40,8 +43,8 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void addCommentToLesson(String userId, String lessonId, CommentDto commentDto) {
-        User user = adminServiceImpl.findUserById(userId);
+    public void addCommentToLesson(String lessonId, CommentDto commentDto) {
+        User user = authHelper.getAuthenticatedUser();
         Lesson lesson = lessonServiceImpl.findLessonById(lessonId);
         Comment comment = Comment.builder()
                 .content(commentDto.getContent())
@@ -74,7 +77,12 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void deleteComment(String commentId) {
-        commentRepository.deleteById(commentId);
+        User user = authHelper.getAuthenticatedUser();
+        Optional<Comment> comment = commentRepository.findCommentByIdAndUser(commentId, user);
+
+        if (comment.isPresent()) {
+            commentRepository.deleteById(commentId);
+        }
     }
 
     public Comment findCommentById(String commentId) {
